+++ tests/test_environment_MatchEnv.py
@@ -41,7 +41,7 @@ class TestCluster:
             match.cluster_environments
 
         query_args = dict(r_guess=r_max, num_neighbors=num_neighbors)
-        match.compute((box, xyz), threshold, neighbors=query_args)
+        match.compute((box, xyz), threshold, cluster_neighbors=query_args)
 
         cluster_env = match.cluster_environments
 
@@ -85,7 +85,7 @@ class TestCluster:
 
         match = freud.environment.EnvironmentCluster()
         query_args = dict(r_guess=r_max, num_neighbors=num_neighbors)
-        match.compute((box, xyz), threshold, neighbors=query_args)
+        match.compute((box, xyz), threshold, cluster_neighbors=query_args)
 
         cluster_env = match.cluster_environments
 
@@ -130,7 +130,9 @@ class TestCluster:
 
         match = freud.environment.EnvironmentCluster()
         query_args = dict(r_guess=r_max, num_neighbors=num_neighbors)
-        match.compute((box, xyz), threshold, registration=False, neighbors=query_args)
+        match.compute(
+            (box, xyz), threshold, registration=False, cluster_neighbors=query_args
+        )
 
         cluster_env = match.cluster_environments
 
@@ -175,7 +177,9 @@ class TestCluster:
 
         match = freud.environment.EnvironmentCluster()
         query_args = dict(r_max=r_max, num_neighbors=num_neighbors)
-        match.compute((box, xyz), threshold, registration=False, neighbors=query_args)
+        match.compute(
+            (box, xyz), threshold, registration=False, cluster_neighbors=query_args
+        )
         cluster_env = match.cluster_environments
 
         fn = os.path.join(self.test_folder, "bcc_env.npy")
@@ -258,7 +262,7 @@ class TestCluster:
             (box, xyz),
             threshold,
             registration=True,
-            neighbors=nlist,
+            cluster_neighbors=nlist,
             env_neighbors=query_args,
         )
         clusters = match.cluster_idx
@@ -437,7 +441,7 @@ class TestCluster:
         assert match._repr_png_() is None
 
         query_args = dict(r_guess=r_max, num_neighbors=num_neighbors)
-        match.compute((box, xyz), threshold, neighbors=query_args)
+        match.compute((box, xyz), threshold, cluster_neighbors=query_args)
         match._repr_png_()
         plt.close("all")
 
@@ -454,7 +458,7 @@ class TestEnvironmentMotifMatch:
         box = freud.box.Box.square(3)
         query_args = dict(r_guess=r_max, num_neighbors=num_neighbors)
         match = freud.environment.EnvironmentMotifMatch().compute(
-            (box, points), motif, 0.1, neighbors=query_args
+            (box, points), motif, 0.1, env_neighbors=query_args
         )
         matches = match.matches
 
@@ -472,7 +476,7 @@ class TestEnvironmentMotifMatch:
         match = freud.environment.EnvironmentMotifMatch()
         query_args = dict(num_neighbors=num_neighbors)
         with pytest.warns(RuntimeWarning):
-            match.compute((box, motif), motif, 0.1, neighbors=query_args)
+            match.compute((box, motif), motif, 0.1, env_neighbors=query_args)
 
 
 class TestEnvironmentRMSDMinimizer: